requirements.txt
fastapi
uvicorn
langchain
langchain-community
openai
qdrant-client[local]
pdfplumber
python-docx
python-dotenv
pydantic
requests
langchain-openai
langgraph
langchain-qdrant
psutil